grader/requirements.txt
Pillow>=10.0.0
numpy>=1.24.0
firebase-admin>=6.2.0
requests>=2.31.0
tqdm>=4.65.0
cloudinary>=1.36.0
anthropic>=0.25.0
